Moving `Image` components into one story category
