+++ stories/Images/FloatImage/floatRight.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { Heading, HugeHeading } from '../../src/atoms/Heading';
-import { DemoContainer } from '../storybook-components';
+import { Heading, HugeHeading } from '../../../src/atoms/Heading';
+import { DemoContainer } from '../../storybook-components';
 
-import { Code } from '../../src/atoms/Code';
+import { Code } from '../../../src/atoms/Code';
 
-import { FloatImage } from '../../src/molecules/FloatImage';
+import { FloatImage } from '../../../src/molecules/FloatImage';
 
 export default () => (
 	<section>
