--- stories/LazyProgressiveImage/index.js
@@ -1,12 +1,0 @@
-import { storiesOf } from '@storybook/react';
-import { StorybookPaddedGrid } from '../../src/storybook-decorators/PaddedGrid';
-
-import LazyProgressiveImage from './basic';
-import MultipleLazyProgressiveImagesStory from './without-blur';
-
-export default () => {
-	storiesOf('LazyProgressiveImage', module)
-		.addDecorator(StorybookPaddedGrid)
-		.add('Low quality image placeholder', LazyProgressiveImage)
-		.add('... without blur transition', MultipleLazyProgressiveImagesStory);
-};
