+++ stories/index.js
@@ -5,14 +5,13 @@ import typography from './typography';
 import editorialMarking from './EditorialMarking';
 import linkBars from './link-bars';
 import trysilPlug from './TrysilPlug';
-import floatImage from './FloatImage';
 import ads from './Ads';
-import lazyProgressiveImage from './LazyProgressiveImage';
 import matComponents from './MatComponents';
 import logos from './logos';
 import forms from './forms';
 import rating from './rating';
 import modals from './Modal';
+import images from './Images';
 
 introduction();
 
@@ -24,12 +23,11 @@ grid();
 rating();
 
 // Molecules
-linkBars();
-trysilPlug();
-floatImage();
 ads();
-lazyProgressiveImage();
-matComponents();
-logos();
 forms();
+images();
+linkBars();
+logos();
+matComponents();
 modals();
+trysilPlug();
